Apply most of @victorskl review comments, data processor under own class for reports
